extract npm package
